ai-service/requirements.txt
flask==3.1.0
flask-limiter==3.9.0
groq==0.11.0
requests==2.32.3
python-dotenv==1.0.1
redis==5.2.1
sentence-transformers==3.3.1
chromadb==0.4.24
gunicorn==23.0.0
pytest==8.3.4
pytest-mock==3.14.0